created etag service and updated app.rb to use it
