+++ lib/meda/collector/app.rb
@@ -14,6 +14,7 @@ require 'meda/services/validation/validation_service'
 require 'meda/services/config/dynamic_config_service'
 require 'meda/services/verification/request_verification_service'
 require 'meda/services/profile/one_key/profile_id_service'
+require 'meda/services/etag/etag_service'
 
 module Meda
   module Collector
@@ -42,6 +43,7 @@ module Meda
 
       @@logging_meta_data_service = Meda::LoggingMetaDataService.new(helper_config)
       @@validation_service = Meda::ValidationService.new
+      @@etag_service = Meda::EtagService.new
       @@dynamic_config_service = Meda::DynamicConfigService.new(Meda.configuration)
       @@request_verification_service = Meda::RequestVerificationService.new(Meda.configuration)
       @@profile_id_service = Meda::ProfileIdService.new(helper_config)
@@ -92,7 +94,7 @@ module Meda
         if Meda.features.is_enabled("etag", false) &&
             Browser.new({:ua => request.env["HTTP_USER_AGENT"].to_s}).safari?
           logger.info("etag is enabled")
-          client_id = get_client_id_from_etag(get_current_etag)
+          client_id = @@etag_service.get_client_id_from_etag(@@etag_service.get_current_etag)
           if client_id.blank?
             client_id = get_client_id_from_cookie
           end
@@ -237,7 +239,7 @@ module Meda
 
           if Meda.features.is_enabled("etag", false) &&
               Browser.new({:ua => request.env["HTTP_USER_AGENT"].to_s}).safari?
-            etag hash_to_string(set_etag_profile_id(profile_id, get_current_etag))
+            etag @@etag_service.hash_to_string(@@etag_service.set_etag_profile_id(profile_id, @@etag_service.get_current_etag))
           end
 
           respond_with_pixel
@@ -246,82 +248,6 @@ module Meda
           logger.error(msg)
           @@request_verification_service.end_rva_log(:status => 'bad_request', :message => msg)
           respond_with_bad_request
-          respond_with_bad_request
-        end
-      end
-
-      def set_etag_profile_id(profile_id, etag_hash)
-        if !etag_hash.blank? && !profile_id.blank?
-          logger.info("about to parse etag string #{etag_hash}")
-          etag_hash = etag_hash
-          etag_hash["profile_id"] = profile_id
-          logger.info("updated etag with profile id #{etag_hash}")
-          return etag_hash
-        else
-          logger.warn("profile_id or etag_hash is empty")
-        end
-      end
-
-      # Get ID from user etag, if
-      def get_current_etag
-        origin_etag = request.env['HTTP_IF_NONE_MATCH'].to_s.clone
-        etag = origin_etag.gsub!(/\A"|"\Z/, '')
-        if etag.blank?
-          new_etag = create_etag_hash
-          logger.info("creating new etag #{new_etag}")
-          return new_etag
-        end
-        logger.info("etag exist in the HTTP_IF_NONE_MATCH header, returning etag: #{etag}")
-        string_to_hash(etag)
-      end
-
-      # Accepts a string in the following format
-      # "client_id=123;profile_id=321;" and converts
-      # it to a ruby hash in the form
-      # { "client_id" => 123, "profile_id", 321}
-      def string_to_hash(str)
-        Hash[
-            str.split(';').map do |pair|
-              k, v = pair.split('=', 2)
-              [k, v]
-            end]
-      end
-
-      # Convert has in the form
-      # { "client_id" => 123, "profile_id", 321}
-      # to string "client_id=123;profile_id=321"
-      def hash_to_string(hash)
-        # TODO lookup more efficient string builder
-        str = ''
-        hash.each do |key, value|
-          str << key.to_s + '=' + value.to_s + ';'
-        end
-        str
-      end
-
-      def create_etag_hash
-        string_to_hash(create_etag_string)
-      end
-
-      def create_etag_string
-        etag = "client_id=#{UUIDTools::UUID.random_create.to_s};profile_id=#{" "};"
-        return etag
-      end
-
-
-      def get_profile_id_from_etag(etag_hash)
-        if !etag_hash.blank? && etag_hash.key?("profile_id")
-          return etag_hash["profile_id"]
-        else
-          logger.warn("tried to get profile id from json_str, but it's empty, or key does not exist")
-        end
-      end
-
-      def get_client_id_from_etag(etag_hash)
-        if !etag_hash.blank? && etag_hash.key?("client_id")
-          return etag_hash["client_id"]
-        else
-          logger.warn("tried to get client_id from json_str, but it's empty, or key does not exist")
         end
       end
 
@@ -432,14 +358,13 @@ module Meda
 
         if Meda.features.is_enabled("etag", false) &&
             Browser.new({:ua => request.env["HTTP_USER_AGENT"].to_s}).safari?
-          profile_id = get_profile_id_from_etag(get_current_etag)
+          profile_id = @@etag_service.get_profile_id_from_etag(@@etag_service.get_current_etag)
           if profile_id.blank?
             profile_id = get_profile_id_from_cookie
           end
         else
           profile_id = get_profile_id_from_cookie
         end
-
 
         params[:profile_id] = profile_id
 
@@ -452,7 +377,7 @@ module Meda
 
           if Meda.features.is_enabled("etag", false) &&
               Browser.new({:ua => request.env["HTTP_USER_AGENT"].to_s}).safari?
-            etag hash_to_string(set_etag_profile_id(params[:profile_id], get_current_etag))
+            etag @@etag_service.hash_to_string(@@etag_service.set_etag_profile_id(params[:profile_id], @@etag_service.get_current_etag))
           end
 
           respond_with_pixel
@@ -521,7 +446,7 @@ module Meda
 
         if Meda.features.is_enabled("etag", false) &&
             Browser.new({:ua => request.env["HTTP_USER_AGENT"].to_s}).safari?
-          profile_id = get_profile_id_from_etag(get_current_etag)
+          profile_id = @@etag_service.get_profile_id_from_etag(@@etag_service.get_current_etag)
           if profile_id.blank?
             profile_id = get_profile_id_from_cookie
           end
@@ -539,9 +464,9 @@ module Meda
           if Meda.features.is_enabled("etag", false) &&
               Browser.new({:ua => request.env["HTTP_USER_AGENT"].to_s}).safari?
               if profile_id.blank?
-                etag hash_to_string(get_current_etag)
+                etag @@etag_service.hash_to_string(@@etag_service.get_current_etag)
               else
-                etag hash_to_string(set_etag_profile_id(profile_id, get_current_etag))
+                etag @@etag_service.hash_to_string(@@etag_service.set_etag_profile_id(profile_id, @@etag_service.get_current_etag))
               end
           end
           respond_with_pixel
@@ -582,7 +507,7 @@ module Meda
 
         if Meda.features.is_enabled("etag", false) &&
             Browser.new({:ua => request.env["HTTP_USER_AGENT"].to_s}).safari?
-          profile_id = get_profile_id_from_etag(get_current_etag)
+          profile_id = @@etag_service.get_profile_id_from_etag(@@etag_service.get_current_etag)
           if profile_id.blank?
             profile_id = get_profile_id_from_cookie
           end
@@ -600,9 +525,9 @@ module Meda
           if Meda.features.is_enabled("etag", false) &&
               Browser.new({:ua => request.env["HTTP_USER_AGENT"].to_s}).safari?
             if profile_id.blank?
-              etag hash_to_string(get_current_etag)
+              etag @@etag_service.hash_to_string(@@etag_service.get_current_etag)
             else
-              etag hash_to_string(set_etag_profile_id(profile_id, get_current_etag))
+              etag @@etag_service.hash_to_string(@@etag_service.set_etag_profile_id(profile_id, @@etag_service.get_current_etag))
             end
           end
           respond_with_pixel
